Crate inject_error

Inject non-deterministic anyhow::Error on annotated functions, it requires the dependency of rand and is only enabled when compiled with feature "inject-error" and not in cargo test.

Attribute Macros

inject_error
Proc macro to specify non-deterministic error injection for functions. Specify the chance with #[inject_error(probability = x)] where 0.0 <= x <= 1.0 When injected, the function would return anyhow!("Injected error") immediately. See the example test.rs for usage.